\name{learningMedicine-package}
\alias{learningMedicine-package}
\alias{learningMedicine}
\docType{package}
\title{Machine Learning Assisted Clinical Medicine Tools}
\description{Machine Learning Assisted Clinical Medicine Tools}
\author{D Fantini}
\examples{
## To make sure that the package was correctly installed, run:
credits()
}
